Update app.py
Added 'open_github_repo' command to the Help menu. This command opens the GitHub repository for this package in the user's default browser. Changed function of the 'open_in_browser' method of the 'help' command to open the user_manual.md file in the GitHub repository if the file is not found in the local files.